+++ docs/source/news/index.rst
@@ -168,7 +168,7 @@ Upcoming Release
         * New filter classes: :class:`~acoular.tprocess.TimeExpAverage`, :class:`~acoular.tprocess.FiltFreqWeight`, :class:`~acoular.tprocess.OctaveFilterBank`
         * Demo script is now part of module (see :doc:`../get_started/index` for usage)
     * New class for processing sound card input: :class:`~acoular.sdinput.SoundDeviceSamplesGenerator`
-    * New class for cumulative averaging: :class:`~acoular.tprocess.TimeCumAverage`
+    * New class for cumulative averaging: :class:`~acoular.process.TimeCumAverage`
     * New grid classes :class:`~acoular.grids.ImportGrid`, :class:`~acoular.grids.LineGrid`, :class:`~acoular.grids.MergeGrid`
     * New versatile integration sector classes :class:`~acoular.grids.RectSector`, :class:`~acoular.grids.CircSector`, :class:`~acoular.grids.PolySector`, :class:`~acoular.grids.ConvexSector`, :class:`~acoular.grids.MultiSector`
     * New class for mixing multiple channels :class:`~acoular.tprocess.ChannelMixer`
@@ -196,7 +196,7 @@ Upcoming Release
 19.11
 ------------
     * Adds new classes for handling rotating data, including detection of trigger signals and interpolation of sensor data for virtual array emulation (:class:`~acoular.tprocess.Trigger`, :class:`~acoular.tprocess.AngleTracker`, :class:`~acoular.tprocess.SpatialInterpolator`, :class:`~acoular.tprocess.SpatialInterpolatorRotation`, :class:`~acoular.tprocess.SpatialInterpolatorConstantRotation`)
-    * Introduces new :class:`~acoular.tprocess.SampleSplitter` class, which allows distribution of data streams 
+    * Introduces new :class:`~acoular.process.SampleSplitter` class, which allows distribution of data streams 
     * Adds new (global) caching options for more flexible cache file handling (e.g. never cache results, always cache, cache read-only). See :class:`~acoular.configuration.config` for information on how to use this.
     * User can choose whether to use h5py or pytables package for handling hdf files. See :class:`~acoular.configuration.config` for information on how to use this.
     * Change: BeamformerGIB behaviour (not calculating sources with eigenvalue of zero)
